% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_DownloadData.R
\name{cl_DownloadData}
\alias{cl_DownloadData}
\title{Download files from pangaea : \preformatted{https://doi.pangaea.de/10.1594/PANGAEA.910898}
otherwise from \preformatted{http://obs-vlfr.fr/Pfunction/}}
\usage{
cl_DownloadData(
  month = 0,
  DownloadGeo = TRUE,
  dirdata = "CoastalLight.d",
  alt = 0,
  stat = "mean"
)
}
\arguments{
\item{month}{: [integer] decimal month (0-12). Default = 0.
\itemize{
  \item 0 indicates the climatology over 21 years (1998-2018).
  \item 1 to 12 indicates the decimal value for a single month (1 = January, ...);
}}

\item{DownloadGeo}{: [logical] download the "geographic" file containing
longitude, latitude, depth and surface area of the pixels, mandatory for retrieving pixel locations
(default = TRUE)}

\item{dirdata}{: [character] The directory where the files are stored (default = "./CoastalLight.d");}

\item{alt}{: [integer] the way of downloading (0, 1 or 2). Default = 0
\itemize{
  \item 0 automatically connect to Pangaea database and download
  \item 1 launch your default browser and connect for manual download to URL \preformatted{
        http://obs-vlfr.fr/Pfunction/
                      }  
  \item 2 launch your default browser and connect for manual download to URL \preformatted{
        https://doi.pangaea.de/10.1594/PANGAEA.910898
                      }  
}
N.B. : if you choose a manual download (alt = 1 or 2), you have to download files in the appropriate directory,
       i.e. the argument \code{dirdata} has no effect.}

\item{stat}{: [character] the statistical value you want for variables "par", "kdpar" and "parbottom";
              choose one among "mean" (mean value),
              "min" (minimum), "max" (maximum), "sd" (standard deviation)
                (default = "mean")
N.B. : "min", "max", "sd", are available only at URL \preformatted{
        http://obs-vlfr.fr/Pfunction/
                      }}
}
\value{
: none
}
\description{
Download files used by \code{cl_GetData()} function.
}
\examples{
cl_DownloadData() # annual climatology over 21 years (1998-2018)
cl_DownloadData(month = 1) # monthly climatology (January) over 21 years (1998-2018)
cl_DownloadData(month = 8) # monthly climatology (August) over 21 years (1998-2018)
cl_DownloadData(month = 8, stat = "sd") # idem, 
but download standard deviation of optical parameters
}
