% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_PlotData.R
\name{cl_PlotData}
\alias{cl_PlotData}
\title{Plots the data returned by function \code{cl_GetData()}
or function \code{cl_Transect()}}
\usage{
cl_PlotData(x, vertices = NULL)
}
\arguments{
\item{x}{: [list] a list returned by function \code{cl_GetData()}}

\item{vertices}{: [list] a polygonal line that represents a transect;
each element of this list is a geographic point,
i.e. a vector of length 2 (longitude, latitude); default is NULL;
this argument is only used if type is "Area" in order to add
a transect to the maps.
N.B. : function \code{cl_Transect()} has an optional argument (\code{plt})
       to plot the transect on the maps.}
}
\value{
\itemize{
\item a rasterstack with all variables if type is "Area"
\item NULL otherwise
}
}
\description{
Functions \code{cl_GetData()} or \code{cl_Transect()} return a list; 
plot differ depending on the field \code{type} of this list :
  \describe{  
    \item{"Point"}{no plot}
    \item{"LonTransect"}{variables are plotted vs longitude}
    \item{"LatTransect"}{variables are plotted vs latitude}
    \item{"Area"}{geographic maps of the variables are produced}
    \item{"Transect"}{variables are plotted vs the distance traveled
                     from the first vertex of the polygonal line}
  }
}
\examples{
## See function \code{cl_GetData()} examples
}
