% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_subregion.R
\name{cl_subregion}
\alias{cl_subregion}
\title{Download a P-function for a subregion}
\usage{
cl_subregion(browse = TRUE)
}
\arguments{
\item{browse}{: [logical] do you want to connect to the server with your default browser ? (default = TRUE).
If FALSE no connection, but the url is returned
         (you may connect to this url, independently of the package).}
}
\value{
: 
[character] the url ("http://obs-vlfr.fr/Pfunction/")
}
\description{
Connects to a server in order to calculate and download a P-function for a subregion of one of the three standard regions
}
\examples{
cl_subregion()
}
